Remove boiler price threshold input number
Peak shaving is not worth it.

Turning the boiler off when energy is 0.5 eur/kWh and turning it on when
it's 0.3 eur saves like a cent a day if we can do this 2 out of 24
hours.

During this time, the Quooker stops functioning. We no longer direct
access to boiling water.

All in all, not worh it.